pyproject.toml
[project]
name = "model-hub"
version = "0.1.0"
description = "Local-first Model Hub for managing ML models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlmodel>=0.0.16",
    "alembic>=1.13.1",
    "arq>=0.25.0",
    "redis>=5.0.3",
    "huggingface-hub>=0.22.2",
    "pydantic-settings>=2.2.1",
    "asyncpg>=0.29.0",
    "betterproto[compiler]>=2.0.0b6",
    "grpcio>=1.62.1",
    "protobuf>=4.25.3",
    "typer>=0.12.3",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
    "opentelemetry-exporter-otlp>=1.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/common", "src/hub_service", "src/worker_service"]

[dependency-groups]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "ruff>=0.3.5",
    "mypy>=1.9.0",
    "types-redis>=4.6.0.20240311",
    "pre-commit>=3.7.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "N"]
ignore = ["E501", "B008"] # Ignore E501 (line too long) and B008 (function call in default arg) for FastAPI Depends

[tool.ruff.lint.per-file-ignores]
"alembic/env.py" = ["E402"] # Ignore import not at top of file for Alembic

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true
